changed the uint value again in writeoutdepthsai

+++ lf_toolbox/LFToolbox0.4/SupportFunctions/LFWriteOutDepthSAI_crv.m
@@ -15,8 +15,8 @@ function LFWriteOutDepthSAI_crv(LF, writeDir)
 
     for iter = pxStart:pxDist:pxEnd
         for jter = pxStart:pxDist:pxEnd
-            img = uint16(squeeze(LF(iter,jter,:,:,:)));
-%             img = uint8(squeeze(LF(iter, jter, :, :, :)));
+            % img = uint16(squeeze(LF(iter,jter,:,:,:)));
+            img = uint8(squeeze(LF(iter, jter, :, :, :)));
             if icounter < 10 && jcounter < 10
                 filename = cat(2, writeDir, '/CRV_0',int2str(icounter),'_0',int2str(jcounter),'.png');
             elseif icounter > 9 && jcounter > 9
